[sky] add support for news.sky.com (closes #13055)
